+++ src/axom/sidre/tests/sidre_mcarray.cpp
@@ -150,7 +150,7 @@ void check_storage(MCArray<T>& v)
     for(axom::IndexType j = 0; j < num_components; ++j)
     {
       EXPECT_EQ(v(i, j), i * num_components + j);
-      EXPECT_EQ(v[i * num_components + j], i * num_components + j);
+      EXPECT_EQ(v.flatIndex(i * num_components + j), i * num_components + j);
       EXPECT_EQ(data_ptr[i * num_components + j], i * num_components + j);
     }
   }
@@ -170,7 +170,7 @@ void check_storage(MCArray<T>& v)
     for(axom::IndexType j = 0; j < num_components; ++j)
     {
       EXPECT_EQ(v(i, j), i * j - 5 * i + 7 * j);
-      EXPECT_EQ(v[i * num_components + j], i * j - 5 * i + 7 * j);
+      EXPECT_EQ(v.flatIndex(i * num_components + j), i * j - 5 * i + 7 * j);
       EXPECT_EQ(data_ptr[i * num_components + j], i * j - 5 * i + 7 * j);
     }
   }